==> notebooks/2.define_the_problem.ipynb <==
- **Task**: Predict the code 1-21
- **Performance**: RMSE since ordered
- **Experience**: A simulation
- **Similar Projects**: Detecting credit card fraud
- **Assumptions**: Nothing major
- **Why**: To memorize the logic
- **Benefits**: ...
- **Handoff**: To self
- **Buy-in**: None
- **Data Available**: Simulations and the features
- **Data I Wish I Had**: None
- **Data to Ignore**: Maybe strings
- **Development Data Format**: CSV
- **Missing Data**: None
- **Anonymize Data**: Already done
- **Change Datatypes**: Not necessary, small file of numbers
- **Test Set Size**: Dev set and test set will be 10,0000 samples each
- **Training Set Sampling Plan**: No sampling.  Training will be approximately 5 million samples


import pandas as pd
df = pd.DataFrame()
dfs = []
# for i in range(1,572):
for i in range(1,150):
    temp_df = pd.read_csv("../data/raw/{}.csv".format(i))
    dfs.append(temp_df)
    print(i)
    if i % 100 == 0:
        print(i)
df = pd.concat(dfs, axis=0)
# df.to_csv("../data/interim/raw_combined.csv")

import pandas as pd
import re
import pandas as pd
from pandas import read_csv
from numpy.random import seed
from sklearn.model_selection import train_test_split

# Set seed - ensures that the datasets are split the same way if re-run
seed(32)

# Common function to improve column names
def camel_to_snake_case(name):
    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
df.columns = [camel_to_snake_case(x) for x in df.columns]
# df.rename(columns={'target2': 'target'}, inplace=True)    # IMPORTANT
# df.rename(columns={'target1': 'bluff_code2'}, inplace=True)    # IMPORTANT

# Drop columns
# df.drop(["time"], axis=1, inplace=True)

# Split datasets
train, temp = train_test_split(df, test_size=0.0035, stratify=df['cat4_flop_bluff_second'])
dev, test = train_test_split(temp, test_size=0.50)

# Write results to files
train.to_csv("../data/interim/train.csv", index=False)
dev.to_csv("../data/interim/dev.csv", index=False)
test.to_csv("../data/interim/test.csv", index=False)

# Print the dataframe shapes and show the rows per target value
print(train.shape, dev.shape, test.shape)
print("\nTrain\n", train['target'].value_counts(), "\n\nDev\n", dev['target'].value_counts(), "\n\nTest\n", test['target'].value_counts())
