==> notebooks/10.Excel Exciting Final Output.ipynb <==
import os
import csv
import pandas as pd
from pandasql import sqldf
from collections import defaultdict

df_extract = pd.read_csv("../data/interim/train.csv")
df_extract = df_extract.drop_duplicates()

range_names = ['LJ RFI', 'HJ RFI', 'CO RFI', 'BN RFI', 'SB RFI']
board_types = ['two_tone'] # rainbow; catboost importance found tones is hardly important, a bit for the river
bluff_code_fields = ["cat4_flop_bluff_second", "cat4_turn_bluff_second", "cat4_river_bluff_second"]
bluff_code_names = ['0 bets', '1 bet', '2 bets']

for board_type in board_types:
    for range_name in range_names:
        for bluff_code_field, bluff_code_name in zip(bluff_code_fields, bluff_code_names):
            ## Extract
            df = df_extract.copy()
            df = df.loc[df['tones'] == (3 if board_type == 'rainbow' else 2 if board_type == 'two_tone' else 1)]
            df = df.loc[df['position'] == range_name]
            df = df.drop_duplicates()
            df.sort_values(by=["rank1", "rank2", "rank3"], inplace=True, ascending=False)
            df = df[['rank1', 'rank2', 'rank3', "cat4_flop_bluff_second", "cat4_turn_bluff_second", "cat4_river_bluff_second"]]
            a = defaultdict(str)
            for index, row in df.iterrows():
                if row['rank1'] == row['rank2'] or row['rank2'] == row['rank3']:
                    continue
                a[(row['rank1'], row['rank2'], row['rank3'])] = row[bluff_code_field]
            with open('../reports/final_results/{}/{}_{}.csv'.format(board_type, range_name, bluff_code_name), mode='w') as fout:
                csv_writer = csv.writer(fout, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)

                csv_writer.writerow(['', '', '', '', '', '', '', '', '', '', '', '', '', '', '', range_name, '','','','','','','','','','','','','','','','','','','','','','','','',''])
                csv_writer.writerow(['', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '','','','','','','','','','','','','','','','','','','','','','','','',''])
                csv_writer.writerow(['', 'A', 'K', 'Q', 'J', 'T', "_9", "_8", "_7", "_6", "_5", "_4", "_3", "_2", '', '', 'K', 'Q', 'J', 'T', "_9", "_8", "_7", "_6", "_5", "_4", "_3", "_2", '', '', 'Q', 'J', 'T', "_9", "_8", "_7", "_6", "_5", "_4", "_3", "_2"])
                csv_writer.writerow(['A', a[(14,14,14)], a[(14,14,13)], a[(14,14,12)], a[(14,14,11)], a[(14,14,10)], a[(14,14,9)], a[(14,14,8)], a[(14,14,7)], a[(14,14,6)], a[(14,14,5)], a[(14,14,4)], a[(14,14,3)], a[(14,14,2)], '', 'K', a[(13,13,13)], a[(13,13,12)], a[(13,13,11)], a[(13,13,10)], a[(13,13,9)], a[(13,13,8)], a[(13,13,7)], a[(13,13,6)], a[(13,13,5)], a[(13,13,4)], a[(13,13,3)], a[(13,13,2)], '', 'Q', a[(12,12,12)], a[(12,12,11)], a[(12,12,10)], a[(12,12,9)], a[(12,12,8)], a[(12,12,7)], a[(12,12,6)], a[(12,12,5)], a[(12,12,4)], a[(12,12,3)], a[(12,12,2)], a[(12,12,1)], a[(12,12,0)]])
                csv_writer.writerow(['K', a[(14,13,14)], a[(14,13,13)], a[(14,13,12)], a[(14,13,11)], a[(14,13,10)], a[(14,13,9)], a[(14,13,8)], a[(14,13,7)], a[(14,13,6)], a[(14,13,5)], a[(14,13,4)], a[(14,13,3)], a[(14,13,2)], '', 'Q', a[(13,12,13)], a[(13,12,12)], a[(13,12,11)], a[(13,12,10)], a[(13,12,9)], a[(13,12,8)], a[(13,12,7)], a[(13,12,6)], a[(13,12,5)], a[(13,12,4)], a[(13,12,3)], a[(13,12,2)], '', 'J', a[(12,11,12)], a[(12,11,11)], a[(12,11,10)], a[(12,11,9)], a[(12,11,8)], a[(12,11,7)], a[(12,11,6)], a[(12,11,5)], a[(12,11,4)], a[(12,11,3)], a[(12,11,2)], a[(12,11,1)], a[(12,11,0)]])
                csv_writer.writerow(['Q', a[(14,12,14)], a[(14,12,13)], a[(14,12,12)], a[(14,12,11)], a[(14,12,10)], a[(14,12,9)], a[(14,12,8)], a[(14,12,7)], a[(14,12,6)], a[(14,12,5)], a[(14,12,4)], a[(14,12,3)], a[(14,12,2)], '', 'J', a[(13,11,13)], a[(13,11,12)], a[(13,11,11)], a[(13,11,10)], a[(13,11,9)], a[(13,11,8)], a[(13,11,7)], a[(13,11,6)], a[(13,11,5)], a[(13,11,4)], a[(13,11,3)], a[(13,11,2)], '', 'T', a[(12,10,12)], a[(12,10,11)], a[(12,10,10)], a[(12,10,9)], a[(12,10,8)], a[(12,10,7)], a[(12,10,6)], a[(12,10,5)], a[(12,10,4)], a[(12,10,3)], a[(12,10,2)], a[(12,10,1)], a[(12,10,0)]])
                csv_writer.writerow(['J', a[(14,11,14)], a[(14,11,13)], a[(14,11,12)], a[(14,11,11)], a[(14,11,10)], a[(14,11,9)], a[(14,11,8)], a[(14,11,7)], a[(14,11,6)], a[(14,11,5)], a[(14,11,4)], a[(14,11,3)], a[(14,11,2)], '', 'T', a[(13,10,13)], a[(13,10,12)], a[(13,10,11)], a[(13,10,10)], a[(13,10,9)], a[(13,10,8)], a[(13,10,7)], a[(13,10,6)], a[(13,10,5)], a[(13,10,4)], a[(13,10,3)], a[(13,10,2)], '', "_9", a[(12,9,12)], a[(12,9,11)], a[(12,9,10)], a[(12,9,9)], a[(12,9,8)], a[(12,9,7)], a[(12,9,6)], a[(12,9,5)], a[(12,9,4)], a[(12,9,3)], a[(12,9,2)], a[(12,9,1)], a[(12,9,0)]])
                csv_writer.writerow(['T', a[(14,10,14)], a[(14,10,13)], a[(14,10,12)], a[(14,10,11)], a[(14,10,10)], a[(14,10,9)], a[(14,10,8)], a[(14,10,7)], a[(14,10,6)], a[(14,10,5)], a[(14,10,4)], a[(14,10,3)], a[(14,10,2)], '', "_9", a[(13,9,13)], a[(13,9,12)], a[(13,9,11)], a[(13,9,10)], a[(13,9,9)], a[(13,9,8)], a[(13,9,7)], a[(13,9,6)], a[(13,9,5)], a[(13,9,4)], a[(13,9,3)], a[(13,9,2)], '', "_8", a[(12,8,12)], a[(12,8,11)], a[(12,8,10)], a[(12,8,9)], a[(12,8,8)], a[(12,8,7)], a[(12,8,6)], a[(12,8,5)], a[(12,8,4)], a[(12,8,3)], a[(12,8,2)], a[(12,8,1)], a[(12,8,0)]])
                csv_writer.writerow(["_9", a[(14,9,14)], a[(14,9,13)], a[(14,9,12)], a[(14,9,11)], a[(14,9,10)], a[(14,9,9)], a[(14,9,8)], a[(14,9,7)], a[(14,9,6)], a[(14,9,5)], a[(14,9,4)], a[(14,9,3)], a[(14,9,2)], '', "_8", a[(13,8,13)], a[(13,8,12)], a[(13,8,11)], a[(13,8,10)], a[(13,8,9)], a[(13,8,8)], a[(13,8,7)], a[(13,8,6)], a[(13,8,5)], a[(13,8,4)], a[(13,8,3)], a[(13,8,2)], '', "_7", a[(12,7,12)], a[(12,7,11)], a[(12,7,10)], a[(12,7,9)], a[(12,7,8)], a[(12,7,7)], a[(12,7,6)], a[(12,7,5)], a[(12,7,4)], a[(12,7,3)], a[(12,7,2)], a[(12,7,1)], a[(12,7,0)]])
                csv_writer.writerow(["_8", a[(14,8,14)], a[(14,8,13)], a[(14,8,12)], a[(14,8,11)], a[(14,8,10)], a[(14,8,9)], a[(14,8,8)], a[(14,8,7)], a[(14,8,6)], a[(14,8,5)], a[(14,8,4)], a[(14,8,3)], a[(14,8,2)], '', "_7", a[(13,7,13)], a[(13,7,12)], a[(13,7,11)], a[(13,7,10)], a[(13,7,9)], a[(13,7,8)], a[(13,7,7)], a[(13,7,6)], a[(13,7,5)], a[(13,7,4)], a[(13,7,3)], a[(13,7,2)], '', "_6", a[(12,6,12)], a[(12,6,11)], a[(12,6,10)], a[(12,6,9)], a[(12,6,8)], a[(12,6,7)], a[(12,6,6)], a[(12,6,5)], a[(12,6,4)], a[(12,6,3)], a[(12,6,2)], a[(12,6,1)], a[(12,6,0)]])
                csv_writer.writerow(["_7", a[(14,7,14)], a[(14,7,13)], a[(14,7,12)], a[(14,7,11)], a[(14,7,10)], a[(14,7,9)], a[(14,7,8)], a[(14,7,7)], a[(14,7,6)], a[(14,7,5)], a[(14,7,4)], a[(14,7,3)], a[(14,7,2)], '', "_6", a[(13,6,13)], a[(13,6,12)], a[(13,6,11)], a[(13,6,10)], a[(13,6,9)], a[(13,6,8)], a[(13,6,7)], a[(13,6,6)], a[(13,6,5)], a[(13,6,4)], a[(13,6,3)], a[(13,6,2)], '', "_5", a[(12,5,12)], a[(12,5,11)], a[(12,5,10)], a[(12,5,9)], a[(12,5,8)], a[(12,5,7)], a[(12,5,6)], a[(12,5,5)], a[(12,5,4)], a[(12,5,3)], a[(12,5,2)], a[(12,5,1)], a[(12,5,0)]])
                csv_writer.writerow(["_6", a[(14,6,14)], a[(14,6,13)], a[(14,6,12)], a[(14,6,11)], a[(14,6,10)], a[(14,6,9)], a[(14,6,8)], a[(14,6,7)], a[(14,6,6)], a[(14,6,5)], a[(14,6,4)], a[(14,6,3)], a[(14,6,2)], '', "_5", a[(13,5,13)], a[(13,5,12)], a[(13,5,11)], a[(13,5,10)], a[(13,5,9)], a[(13,5,8)], a[(13,5,7)], a[(13,5,6)], a[(13,5,5)], a[(13,5,4)], a[(13,5,3)], a[(13,5,2)], '', "_4", a[(12,4,12)], a[(12,4,11)], a[(12,4,10)], a[(12,4,9)], a[(12,4,8)], a[(12,4,7)], a[(12,4,6)], a[(12,4,5)], a[(12,4,4)], a[(12,4,3)], a[(12,4,2)], a[(12,4,1)], a[(12,4,0)]])
                csv_writer.writerow(["_5", a[(14,5,14)], a[(14,5,13)], a[(14,5,12)], a[(14,5,11)], a[(14,5,10)], a[(14,5,9)], a[(14,5,8)], a[(14,5,7)], a[(14,5,6)], a[(14,5,5)], a[(14,5,4)], a[(14,5,3)], a[(14,5,2)], '', "_4", a[(13,4,13)], a[(13,4,12)], a[(13,4,11)], a[(13,4,10)], a[(13,4,9)], a[(13,4,8)], a[(13,4,7)], a[(13,4,6)], a[(13,4,5)], a[(13,4,4)], a[(13,4,3)], a[(13,4,2)], '', "_3", a[(12,3,12)], a[(12,3,11)], a[(12,3,10)], a[(12,3,9)], a[(12,3,8)], a[(12,3,7)], a[(12,3,6)], a[(12,3,5)], a[(12,3,4)], a[(12,3,3)], a[(12,3,2)], a[(12,3,1)], a[(12,3,0)]])
                csv_writer.writerow(["_4", a[(14,4,14)], a[(14,4,13)], a[(14,4,12)], a[(14,4,11)], a[(14,4,10)], a[(14,4,9)], a[(14,4,8)], a[(14,4,7)], a[(14,4,6)], a[(14,4,5)], a[(14,4,4)], a[(14,4,3)], a[(14,4,2)], '', "_3", a[(13,3,13)], a[(13,3,12)], a[(13,3,11)], a[(13,3,10)], a[(13,3,9)], a[(13,3,8)], a[(13,3,7)], a[(13,3,6)], a[(13,3,5)], a[(13,3,4)], a[(13,3,3)], a[(13,3,2)], '', "_2", a[(12,2,12)], a[(12,2,11)], a[(12,2,10)], a[(12,2,9)], a[(12,2,8)], a[(12,2,7)], a[(12,2,6)], a[(12,2,5)], a[(12,2,4)], a[(12,2,3)], a[(12,2,2)], a[(12,2,1)], a[(12,2,0)]])
                csv_writer.writerow(["_3", a[(14,3,14)], a[(14,3,13)], a[(14,3,12)], a[(14,3,11)], a[(14,3,10)], a[(14,3,9)], a[(14,3,8)], a[(14,3,7)], a[(14,3,6)], a[(14,3,5)], a[(14,3,4)], a[(14,3,3)], a[(14,3,2)], '', "_2", a[(13,2,13)], a[(13,2,12)], a[(13,2,11)], a[(13,2,10)], a[(13,2,9)], a[(13,2,8)], a[(13,2,7)], a[(13,2,6)], a[(13,2,5)], a[(13,2,4)], a[(13,2,3)], a[(13,2,2)], '', '', a[(12,1,12)], a[(12,1,11)], a[(12,1,10)], a[(12,1,9)], a[(12,1,8)], a[(12,1,7)], a[(12,1,6)], a[(12,1,5)], a[(12,1,4)], a[(12,1,3)], a[(12,1,2)], a[(12,1,1)], a[(12,1,0)]])
                csv_writer.writerow(["_2", a[(14,2,14)], a[(14,2,13)], a[(14,2,12)], a[(14,2,11)], a[(14,2,10)], a[(14,2,9)], a[(14,2,8)], a[(14,2,7)], a[(14,2,6)], a[(14,2,5)], a[(14,2,4)], a[(14,2,3)], a[(14,2,2)], '', '', a[(13,1,13)], a[(13,1,12)], a[(13,1,11)], a[(13,1,10)], a[(13,1,9)], a[(13,1,8)], a[(13,1,7)], a[(13,1,6)], a[(13,1,5)], a[(13,1,4)], a[(13,1,3)], a[(13,1,2)], '', '', a[(12,0,12)], a[(12,0,11)], a[(12,0,10)], a[(12,0,9)], a[(12,0,8)], a[(12,0,7)], a[(12,0,6)], a[(12,0,5)], a[(12,0,4)], a[(12,0,3)], a[(12,0,2)], a[(12,0,1)], a[(12,0,0)]])
                csv_writer.writerow(['', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '','','','','','','','','','','','','','','','','','','','','','','','',''])    
                csv_writer.writerow(['', 'J', 'T', "_9", "_8", "_7", "_6", "_5", "_4", "_3", "_2", '', '', 'T', "_9", "_8", "_7", "_6", "_5", "_4", "_3", "_2", '', '', "_9", "_8", "_7", "_6", "_5", "_4", "_3", "_2", '', '', "_8", "_7", "_6", "_5", "_4", "_3", "_2"])    
                csv_writer.writerow(['J', a[(11,11,11)], a[(11,11,10)], a[(11,11,9)], a[(11,11,8)], a[(11,11,7)], a[(11,11,6)], a[(11,11,5)], a[(11,11,4)], a[(11,11,3)], a[(11,11,2)], '', 'T', a[(10,10,10)], a[(10,10,9)], a[(10,10,8)], a[(10,10,7)], a[(10,10,6)], a[(10,10,5)], a[(10,10,4)], a[(10,10,3)], a[(10,10,2)], '', "_9", a[(9,9,9)], a[(9,9,8)], a[(9,9,7)], a[(9,9,6)], a[(9,9,5)], a[(9,9,4)], a[(9,9,3)], a[(9,9,2)], '', "_8", a[(8,8,8)], a[(8,8,7)], a[(8,8,6)], a[(8,8,5)], a[(8,8,4)], a[(8,8,3)], a[(8,8,2)]])
                csv_writer.writerow(['T', a[(11,10,11)], a[(11,10,10)], a[(11,10,9)], a[(11,10,8)], a[(11,10,7)], a[(11,10,6)], a[(11,10,5)], a[(11,10,4)], a[(11,10,3)], a[(11,10,2)], '', "_9", a[(10,9,10)], a[(10,9,9)], a[(10,9,8)], a[(10,9,7)], a[(10,9,6)], a[(10,9,5)], a[(10,9,4)], a[(10,9,3)], a[(10,9,2)], '', "_8", a[(9,8,9)], a[(9,8,8)], a[(9,8,7)], a[(9,8,6)], a[(9,8,5)], a[(9,8,4)], a[(9,8,3)], a[(9,8,2)], '', "_7", a[(8,7,8)], a[(8,7,7)], a[(8,7,6)], a[(8,7,5)], a[(8,7,4)], a[(8,7,3)], a[(8,7,2)]])
                csv_writer.writerow(["_9", a[(11,9,11)], a[(11,9,10)], a[(11,9,9)], a[(11,9,8)], a[(11,9,7)], a[(11,9,6)], a[(11,9,5)], a[(11,9,4)], a[(11,9,3)], a[(11,9,2)], '', "_8", a[(10,8,10)], a[(10,8,9)], a[(10,8,8)], a[(10,8,7)], a[(10,8,6)], a[(10,8,5)], a[(10,8,4)], a[(10,8,3)], a[(10,8,2)], '', "_7", a[(9,7,9)], a[(9,7,8)], a[(9,7,7)], a[(9,7,6)], a[(9,7,5)], a[(9,7,4)], a[(9,7,3)], a[(9,7,2)], '', "_6", a[(8,6,8)], a[(8,6,7)], a[(8,6,6)], a[(8,6,5)], a[(8,6,4)], a[(8,6,3)], a[(8,6,2)]])
                csv_writer.writerow(["_8", a[(11,8,11)], a[(11,8,10)], a[(11,8,9)], a[(11,8,8)], a[(11,8,7)], a[(11,8,6)], a[(11,8,5)], a[(11,8,4)], a[(11,8,3)], a[(11,8,2)], '', "_7", a[(10,7,10)], a[(10,7,9)], a[(10,7,8)], a[(10,7,7)], a[(10,7,6)], a[(10,7,5)], a[(10,7,4)], a[(10,7,3)], a[(10,7,2)], '', "_6", a[(9,6,9)], a[(9,6,8)], a[(9,6,7)], a[(9,6,6)], a[(9,6,5)], a[(9,6,4)], a[(9,6,3)], a[(9,6,2)], '', "_5", a[(8,5,8)], a[(8,5,7)], a[(8,5,6)], a[(8,5,5)], a[(8,5,4)], a[(8,5,3)], a[(8,5,2)]])
                csv_writer.writerow(["_7", a[(11,7,11)], a[(11,7,10)], a[(11,7,9)], a[(11,7,8)], a[(11,7,7)], a[(11,7,6)], a[(11,7,5)], a[(11,7,4)], a[(11,7,3)], a[(11,7,2)], '', "_6", a[(10,6,10)], a[(10,6,9)], a[(10,6,8)], a[(10,6,7)], a[(10,6,6)], a[(10,6,5)], a[(10,6,4)], a[(10,6,3)], a[(10,6,2)], '', "_5", a[(9,5,9)], a[(9,5,8)], a[(9,5,7)], a[(9,5,6)], a[(9,5,5)], a[(9,5,4)], a[(9,5,3)], a[(9,5,2)], '', "_4", a[(8,4,8)], a[(8,4,7)], a[(8,4,6)], a[(8,4,5)], a[(8,4,4)], a[(8,4,3)], a[(8,4,2)]])
                csv_writer.writerow(["_6", a[(11,6,11)], a[(11,6,10)], a[(11,6,9)], a[(11,6,8)], a[(11,6,7)], a[(11,6,6)], a[(11,6,5)], a[(11,6,4)], a[(11,6,3)], a[(11,6,2)], '', "_5", a[(10,5,10)], a[(10,5,9)], a[(10,5,8)], a[(10,5,7)], a[(10,5,6)], a[(10,5,5)], a[(10,5,4)], a[(10,5,3)], a[(10,5,2)], '', "_4", a[(9,4,9)], a[(9,4,8)], a[(9,4,7)], a[(9,4,6)], a[(9,4,5)], a[(9,4,4)], a[(9,4,3)], a[(9,4,2)], '', "_3", a[(8,3,8)], a[(8,3,7)], a[(8,3,6)], a[(8,3,5)], a[(8,3,4)], a[(8,3,3)], a[(8,3,2)]])
                csv_writer.writerow(["_5", a[(11,5,11)], a[(11,5,10)], a[(11,5,9)], a[(11,5,8)], a[(11,5,7)], a[(11,5,6)], a[(11,5,5)], a[(11,5,4)], a[(11,5,3)], a[(11,5,2)], '', "_4", a[(10,4,10)], a[(10,4,9)], a[(10,4,8)], a[(10,4,7)], a[(10,4,6)], a[(10,4,5)], a[(10,4,4)], a[(10,4,3)], a[(10,4,2)], '', "_3", a[(9,3,9)], a[(9,3,8)], a[(9,3,7)], a[(9,3,6)], a[(9,3,5)], a[(9,3,4)], a[(9,3,3)], a[(9,3,2)], '', "_2", a[(8,2,8)], a[(8,2,7)], a[(8,2,6)], a[(8,2,5)], a[(8,2,4)], a[(8,2,3)], a[(8,2,2)]])
                csv_writer.writerow(["_4", a[(11,4,11)], a[(11,4,10)], a[(11,4,9)], a[(11,4,8)], a[(11,4,7)], a[(11,4,6)], a[(11,4,5)], a[(11,4,4)], a[(11,4,3)], a[(11,4,2)], '', "_3", a[(10,3,10)], a[(10,3,9)], a[(10,3,8)], a[(10,3,7)], a[(10,3,6)], a[(10,3,5)], a[(10,3,4)], a[(10,3,3)], a[(10,3,2)], '', "_2", a[(9,2,9)], a[(9,2,8)], a[(9,2,7)], a[(9,2,6)], a[(9,2,5)], a[(9,2,4)], a[(9,2,3)], a[(9,2,2)], '', '', a[(8,1,8)], a[(8,1,7)], a[(8,1,6)], a[(8,1,5)], a[(8,1,4)], a[(8,1,3)], a[(8,1,2)]])
                csv_writer.writerow(["_3", a[(11,3,11)], a[(11,3,10)], a[(11,3,9)], a[(11,3,8)], a[(11,3,7)], a[(11,3,6)], a[(11,3,5)], a[(11,3,4)], a[(11,3,3)], a[(11,3,2)], '', "_2", a[(10,2,10)], a[(10,2,9)], a[(10,2,8)], a[(10,2,7)], a[(10,2,6)], a[(10,2,5)], a[(10,2,4)], a[(10,2,3)], a[(10,2,2)], '', '', a[(9,1,9)], a[(9,1,8)], a[(9,1,7)], a[(9,1,6)], a[(9,1,5)], a[(9,1,4)], a[(9,1,3)], a[(9,1,2)], '', '', a[(8,0,8)], a[(8,0,7)], a[(8,0,6)], a[(8,0,5)], a[(8,0,4)], a[(8,0,3)], a[(8,0,2)]])
                csv_writer.writerow(["_2", a[(11,2,11)], a[(11,2,10)], a[(11,2,9)], a[(11,2,8)], a[(11,2,7)], a[(11,2,6)], a[(11,2,5)], a[(11,2,4)], a[(11,2,3)], a[(11,2,2)], '', '', a[(10,1,10)], a[(10,1,9)], a[(10,1,8)], a[(10,1,7)], a[(10,1,6)], a[(10,1,5)], a[(10,1,4)], a[(10,1,3)], a[(10,1,2)], '', '', a[(9,0,9)], a[(9,0,8)], a[(9,0,7)], a[(9,0,6)], a[(9,0,5)], a[(9,0,4)], a[(9,0,3)], a[(9,0,2)], '', '', a[(8,-1,8)], a[(8,-1,7)], a[(8,-1,6)], a[(8,-1,5)], a[(8,-1,4)], a[(8,-1,3)], a[(8,-1,2)]])
                csv_writer.writerow(['', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '', '','','','','','','','','','','','','','','','','','','','','','','','',''])    
                csv_writer.writerow(['', "_7", "_6", "_5", "_4", "_3", "_2", '', '', "_6", "_5", "_4", "_3", "_2", '', '', "_5", "_4", "_3", "_2", '', '', "_4", "_3", "_2", '', '', "_3", "_2", '', '', "_2"])
                csv_writer.writerow(["_7", a[(7,7,7)], a[(7,7,6)], a[(7,7,5)], a[(7,7,4)], a[(7,7,3)], a[(7,7,2)], '', "_6", a[(6,6,6)], a[(6,6,5)], a[(6,6,4)], a[(6,6,3)], a[(6,6,2)], '', "_5", a[(5,5,5)], a[(5,5,4)], a[(5,5,3)], a[(5,5,2)], '', "_4", a[(4,4,4)], a[(4,4,3)], a[(4,4,2)], '', "_3", a[(3,3,3)], a[(3,3,2)], '', "_2", a[(2,2,2)]])
                csv_writer.writerow(["_6", a[(7,6,7)], a[(7,6,6)], a[(7,6,5)], a[(7,6,4)], a[(7,6,3)], a[(7,6,2)], '', "_5", a[(6,5,6)], a[(6,5,5)], a[(6,5,4)], a[(6,5,3)], a[(6,5,2)], '', "_4", a[(5,4,5)], a[(5,4,4)], a[(5,4,3)], a[(5,4,2)], '', "_3", a[(4,3,4)], a[(4,3,3)], a[(4,3,2)], '', "_2", a[(3,2,3)], a[(3,2,2)], '', '', a[(2,1,2)]])
                csv_writer.writerow(["_5", a[(7,5,7)], a[(7,5,6)], a[(7,5,5)], a[(7,5,4)], a[(7,5,3)], a[(7,5,2)], '', "_4", a[(6,4,6)], a[(6,4,5)], a[(6,4,4)], a[(6,4,3)], a[(6,4,2)], '', "_3", a[(5,3,5)], a[(5,3,4)], a[(5,3,3)], a[(5,3,2)], '', "_2", a[(4,2,4)], a[(4,2,3)], a[(4,2,2)], '', '', a[(3,1,3)], a[(3,1,2)], '', '', a[(2,0,2)]])
                csv_writer.writerow(["_4", a[(7,4,7)], a[(7,4,6)], a[(7,4,5)], a[(7,4,4)], a[(7,4,3)], a[(7,4,2)], '', "_3", a[(6,3,6)], a[(6,3,5)], a[(6,3,4)], a[(6,3,3)], a[(6,3,2)], '', "_2", a[(5,2,5)], a[(5,2,4)], a[(5,2,3)], a[(5,2,2)], '', '', a[(4,1,4)], a[(4,1,3)], a[(4,1,2)], '', '', a[(3,0,3)], a[(3,0,2)], '', '', a[(2,-1,2)]])
                csv_writer.writerow(["_3", a[(7,3,7)], a[(7,3,6)], a[(7,3,5)], a[(7,3,4)], a[(7,3,3)], a[(7,3,2)], '', "_2", a[(6,2,6)], a[(6,2,5)], a[(6,2,4)], a[(6,2,3)], a[(6,2,2)], '', '', a[(5,1,5)], a[(5,1,4)], a[(5,1,3)], a[(5,1,2)], '', '', a[(4,0,4)], a[(4,0,3)], a[(4,0,2)], '', '', a[(3,-1,3)], a[(3,-1,2)], '', '', a[(2,-2,2)]])
                csv_writer.writerow(["_2", a[(7,2,7)], a[(7,2,6)], a[(7,2,5)], a[(7,2,4)], a[(7,2,3)], a[(7,2,2)], '', '', a[(6,1,6)], a[(6,1,5)], a[(6,1,4)], a[(6,1,3)], a[(6,1,2)], '', '', a[(5,0,5)], a[(5,0,4)], a[(5,0,3)], a[(5,0,2)], '', '', a[(4,-1,4)], a[(4,-1,3)], a[(4,-1,2)], '', '', a[(3,-2,3)], a[(3,-2,2)], '', '', a[(2,-3,2)]])
